// src/app/model/product.ts
export class CardModel {

    // constructor(
    //     id: number,
    //     Name: string,
    //     Type: string,
    //     APRFrom: number,
    //     APRTo: number,
    //     Description: string,
    //     CreditLimit: number,
    //     LaunchDate: string,
    //     Availability: string) {
    //     this.id = id;
    //     this.Name = Name;
    //     this.Type = Type;
    //     this.APRFrom = APRFrom;
    //     this.APRTo = APRTo;
    //     this.Description = Description;
    //     this.CreditLimit = CreditLimit;
    //     this.LaunchDate = LaunchDate;
    //     this.Availability = Availability;
    // }

    id: number;
    Name: string;
    Type: string;
    APRFrom: number;
    APRTo: number;
    Description: string;
    CreditLimit: number;
    LaunchDate: string;
    Availability: string;
}